--- src/main/java/com/EmployeeManagementSystem/Repository/DepartmentRepository.java
@@ -1,9 +1,0 @@
-package com.EmployeeManagementSystem.Repository;
-
-import com.EmployeeManagementSystem.Entity.Department;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface DepartmentRepository extends JpaRepository<Department,Integer> {
-}
