--- src/main/java/com/EmployeeManagementSystem/Repository/EmployeeRepository.java
@@ -1,9 +1,0 @@
-package com.EmployeeManagementSystem.Repository;
-
-import com.EmployeeManagementSystem.Entity.Employee;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface EmployeeRepository extends JpaRepository<Employee,Integer> {
-}
