+++ src/main/java/com/EmployeeManagementSystem/controller/EmployeeController.java
@@ -1,8 +1,8 @@
-package com.EmployeeManagementSystem.Controller;
+package com.EmployeeManagementSystem.controller;
 
-import com.EmployeeManagementSystem.ReuestDTO.EmployeeDTO;
-import com.EmployeeManagementSystem.Entity.Employee;
-import com.EmployeeManagementSystem.Service.EmployeeService;
+import com.EmployeeManagementSystem.entity.Employee;
+import com.EmployeeManagementSystem.requests.EmployeeDTO;
+import com.EmployeeManagementSystem.service.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,31 +18,31 @@ public class EmployeeController {
     private EmployeeService employeeService;
 
     @PostMapping("/Save")
-    public ResponseEntity<String> saveEmployee(@RequestBody EmployeeDTO employeeDTO){
+    public ResponseEntity<String> saveEmployee(@RequestBody EmployeeDTO employeeDTO) {
         employeeService.addEmployee(employeeDTO);
         return ResponseEntity.ok("New Employee Added Successfully!");
     }
 
     @GetMapping("/GetAll")
-    public List<Employee> getAllEmployee(){
+    public List<Employee> getAllEmployee() {
         return employeeService.getAllEmployee();
     }
 
     @GetMapping("/Get/{employeeId}")
-    public ResponseEntity<Employee> getEmployeeById(@PathVariable Integer employeeId){
-        return new ResponseEntity<Employee>(employeeService.getEmployeeById(employeeId),HttpStatus.FOUND);
+    public ResponseEntity<Employee> getEmployeeById(@PathVariable Integer employeeId) {
+        return new ResponseEntity<Employee>(employeeService.getEmployeeById(employeeId), HttpStatus.FOUND);
     }
 
     @PutMapping("/update/{id}")
     public ResponseEntity<String> updateEmployeeById(@PathVariable Integer id,
-                                                     @RequestBody EmployeeDTO employeeDTO){
+                                                     @RequestBody EmployeeDTO employeeDTO) {
         employeeService.updateEmployeeById(id, employeeDTO);
         return ResponseEntity.ok("Employee details are updated successfully!");
     }
 
     @DeleteMapping("/delete/{employeeId}")
-    public ResponseEntity<String> deleteEmployeeById(@PathVariable Integer employeeId){
+    public ResponseEntity<String> deleteEmployeeById(@PathVariable Integer employeeId) {
         employeeService.deleteEmployeeById(employeeId);
-        return ResponseEntity.ok("Employee id "+employeeId+" is deleted!");
+        return ResponseEntity.ok("Employee id " + employeeId + " is deleted!");
     }
 }
