+++ src/main/java/com/EmployeeManagementSystem/repository/DepartmentRepository.java
@@ -1,0 +1,9 @@
+package com.EmployeeManagementSystem.repository;
+
+import com.EmployeeManagementSystem.entity.Department;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface DepartmentRepository extends JpaRepository<Department, Integer> {
+}
