+++ src/main/java/com/EmployeeManagementSystem/repository/EmployeeRepository.java
@@ -1,0 +1,9 @@
+package com.EmployeeManagementSystem.repository;
+
+import com.EmployeeManagementSystem.entity.Employee;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+}
