+++ src/main/java/com/EmployeeManagementSystem/service/DepartmentService.java
@@ -1,15 +1,19 @@
-package com.EmployeeManagementSystem.Service;
+package com.EmployeeManagementSystem.service;
 
-import com.EmployeeManagementSystem.Entity.Department;
-import com.EmployeeManagementSystem.ReuestDTO.DepartmentDTO;
+import com.EmployeeManagementSystem.entity.Department;
+import com.EmployeeManagementSystem.requests.DepartmentDTO;
 
 import java.util.List;
 
 public interface DepartmentService {
 
     Department addDepartment(DepartmentDTO departmentDTO);
+
     List<Department> getAllDepartment();
+
     Department getDepartmentById(Integer id);
+
     Department updateDepartmentById(Integer id, DepartmentDTO departmentDTO);
+
     void deleteDepartmentById(Integer id);
 }
