+++ src/main/java/com/EmployeeManagementSystem/service/DesignationService.java
@@ -1,14 +1,19 @@
-package com.EmployeeManagementSystem.Service;
+package com.EmployeeManagementSystem.service;
 
-import com.EmployeeManagementSystem.Entity.Designation;
-import com.EmployeeManagementSystem.ReuestDTO.DesignationDTO;
+import com.EmployeeManagementSystem.entity.Designation;
+import com.EmployeeManagementSystem.requests.DesignationDTO;
+
 import java.util.List;
 
 public interface DesignationService {
 
     Designation addDesignation(DesignationDTO designationDTO);
+
     List<Designation> getAllDesignations();
+
     Designation getDesignationById(Integer id);
+
     Designation updateDesignationByd(Integer id, DesignationDTO designationDTO);
+
     void deleteDesignationById(Integer id);
 }
