+++ src/main/java/com/EmployeeManagementSystem/service/EmployeeService.java
@@ -1,15 +1,20 @@
-package com.EmployeeManagementSystem.Service;
+package com.EmployeeManagementSystem.service;
 
-import com.EmployeeManagementSystem.ReuestDTO.EmployeeDTO;
-import com.EmployeeManagementSystem.Entity.Employee;
+import com.EmployeeManagementSystem.entity.Employee;
+import com.EmployeeManagementSystem.requests.EmployeeDTO;
+
 import java.util.List;
 
 
 public interface EmployeeService {
     Employee addEmployee(EmployeeDTO employeeDTO);
+
     List<Employee> getAllEmployee();
+
     Employee getEmployeeById(Integer id);
+
     Employee updateEmployeeById(Integer id, EmployeeDTO employeeDTO);
+
     void deleteEmployeeById(Integer employeeId);
